requirements.txt
# Wear Prediction - IoT Cargo Logistics
# Python 3.8+

numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
xgboost>=1.5.0
matplotlib>=3.5.0
seaborn>=0.11.0
scipy>=1.7.0
tensorflow>=2.8.0

# Backend (Flask) + DB
flask>=2.0.0
flask-cors>=3.0.0
mysql-connector-python>=8.0.0
python-dotenv>=1.0.0

# Frontend (Streamlit)
streamlit>=1.20.0
requests>=2.26.0
